feat(press): /press-releases index with draft/published statuses
Turn the single /press-release page into a press room: an index listing
multiple releases, each with a draft|published status. Drafts are hidden
from the public list and marked noindex, but reachable by direct link
(and shown in the index with ?preview=1) for review before publishing.

- src/lib/press-releases.ts: release registry (metadata + status) and helpers
- /press-releases: public index (published only; drafts via ?preview=1)
- /press-releases/[slug]: per-release detail, noindex + banner for drafts
- Release bodies extracted to src/components/press/releases/
- New June 2026 launch release (draft) with live-verified numbers
  (13,000+ translated, ~4M pages) and an English Wikipedia scale comparison
- Legacy /press-release now redirects to the newest published release
